Fix server: don't use spdy on node >= v10.0.0
Ports fix for https://github.com/webpack/webpack-dev-server/issues/1449 to v2 branch
cherry pick of https://github.com/webpack/webpack-dev-server/pull/1451/commits/e97d345ac370095a6e339b7997b939c88ef3e81b

this issue was fixed in webpack-dev-server v3 line, which requires webpack to be
upgraded to >= 4.0.0. This commit cherry picks the fix to v2 branch (on top of v2.11.3)
which does not require webpack to be upgraded to 4.0.0

`spdy` is effectively unmaintained, and as a consequence of an
implementation that extensively relies on Node’s non-public APIs, broken
on Node 10 and above. In those cases, only https will be used for now.
Once express supports Node's built-in HTTP/2 support, migrating over to
that should be the best way to go.

related issues:
nodejs/node#21665
https://github.com/nodejs/node/issues/21665
https://github.com/webpack/webpack-dev-server/issues/1449
https://github.com/expressjs/express/issues/3388